astromesh-orbit/pyproject.toml
[project]
name = "astromesh-orbit"
version = "0.1.0"
description = "Cloud-native deployment for Astromesh — provision infrastructure on any cloud"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
dependencies = [
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "typer>=0.12",
]

[project.optional-dependencies]
gcp = [
    "google-cloud-resource-manager>=1.12",
    "google-auth>=2.29",
]
aws = []
azure = []
all = ["astromesh-orbit[gcp]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.entry-points."astromeshctl.plugins"]
orbit = "astromesh_orbit.cli:register"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["astromesh_orbit"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "orbit_gcp: tests that create real GCP resources (deselect with -m 'not orbit_gcp')",
]
